usersApi and apiClient refactor for better routing, zodios folder restructuring
