+++ website/zodios/apiClient.ts
@@ -1,11 +1,16 @@
-import { Zodios } from "@zodios/core";
+import { Zodios, mergeApis } from "@zodios/core";
 
-import API_BASE_URL from "@/zodios/baseURL";
 import usersApi from "@/zodios/users/usersApi";
-import withCredentialsPlugin from "@/zodios/withCredentialsPlugin";
-import loggerPlugin from "@/zodios/loggerPlugin";
+import API_BASE_URL from "@/zodios/utilities/baseURL";
 
-const apiClient = new Zodios(API_BASE_URL, usersApi);
+import withCredentialsPlugin from "@/zodios/plugins/withCredentialsPlugin";
+import loggerPlugin from "@/zodios/plugins/loggerPlugin";
+
+const api = mergeApis({
+    "/users": usersApi,
+});
+
+const apiClient = new Zodios(API_BASE_URL, api);
 apiClient.use(withCredentialsPlugin);
 apiClient.use(loggerPlugin);
